Tensorflow/requirements.txt
# tensorflow-gpu
tensorflow-hub
tensorflow-datasets
tensorflow_text
pandas
numpy
matplotlib
protobuf==3.20
